Move sharing related files to the ContentRepository/Sharing
