+++ src/Search/Querying/SnQueryController.cs
@@ -125,29 +125,30 @@ namespace SenseNet.Search.Querying
 
         internal static SnQuery ApplyVisitors(SnQuery query) //UNDONE:<? New extensibility pont: query rewriter can be a plugin.
         {
-            var queryTree = query.QueryTree;
+throw new NotImplementedException();
+            //var queryTree = query.QueryTree;
 
-            var visitedTree = new SharingVisitor().Visit(queryTree);
-            if (ReferenceEquals(visitedTree, queryTree))
-                return query;
+            //var visitedTree = new SharingVisitor().Visit(queryTree);
+            //if (ReferenceEquals(visitedTree, queryTree))
+            //    return query;
 
-            var newQuery = Create(visitedTree);
+            //var newQuery = Create(visitedTree);
 
-            newQuery.Querytext = query.Querytext;
-            newQuery.Projection = query.Projection;
-            newQuery.Top = query.Top;
-            newQuery.Skip = query.Skip;
-            newQuery.Sort = query.Sort;
-            newQuery.EnableAutofilters = query.EnableAutofilters;
-            newQuery.EnableLifespanFilter = query.EnableLifespanFilter;
-            newQuery.CountOnly = query.CountOnly;
-            newQuery.QueryExecutionMode = query.QueryExecutionMode;
-            newQuery.AllVersions = query.AllVersions;
-            newQuery.CountAllPages = query.CountAllPages;
-            newQuery.ThrowIfEmpty = query.ThrowIfEmpty;
-            newQuery.ExistenceOnly = query.ExistenceOnly;
+            //newQuery.Querytext = query.Querytext;
+            //newQuery.Projection = query.Projection;
+            //newQuery.Top = query.Top;
+            //newQuery.Skip = query.Skip;
+            //newQuery.Sort = query.Sort;
+            //newQuery.EnableAutofilters = query.EnableAutofilters;
+            //newQuery.EnableLifespanFilter = query.EnableLifespanFilter;
+            //newQuery.CountOnly = query.CountOnly;
+            //newQuery.QueryExecutionMode = query.QueryExecutionMode;
+            //newQuery.AllVersions = query.AllVersions;
+            //newQuery.CountAllPages = query.CountAllPages;
+            //newQuery.ThrowIfEmpty = query.ThrowIfEmpty;
+            //newQuery.ExistenceOnly = query.ExistenceOnly;
 
-            return newQuery;
+            //return newQuery;
         }
 
         private static bool IsAutofilterEnabled(FilterStatus value)
